pyproject.toml
[project]
name = "streak-gen"
version = "0.1.0"
description = "Generate stained-glass style segmented letters/words as printable SVG."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "numpy>=2.0.0",
  "scipy>=1.12.0",
  "shapely>=2.0.5",
  "pyclipper>=1.3.0.post6",
  "svgwrite>=1.4.3",
  "skia-python>=87.5",
]

[project.scripts]
streak-gen = "streak_gen.cli:app"

[tool.uv]
package = true

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"
